backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
bcrypt
pyjwt
python-dotenv
pydantic[email]